Connection to Bluetooth is now stabilished using Service. To Do: Broadcast and Receive between MainActivity and Service
