+++ app/src/main/AndroidManifest.xml
@@ -22,5 +22,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <service
+            android:name=".BluetoothConnection"
+            android:exported="false"/>
     </application>
 </manifest>
