+++ app/src/main/res/SHAs.txt
@@ -1,0 +1,2 @@
+Antiga SHA1 - 19:0B:E8:CA:FB:56:6C:A1:61:46:86:ED:86:A0:CB:7D:54:98:45:51
+Atual SHA1- 2C:DA:B9:9E:B7:7E:D6:51:69:2E:98:91:DA:44:06:AD:AA:D0:25:E2
